% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{ac}
\alias{ac}
\title{Autocorrelation and partial autocorrealtion function.}
\usage{
ac(series, max.lag = 18, main = NULL)
}
\arguments{
\item{series}{A numeric, time series, or xts variable.}

\item{max.lag}{A number that represents the maximum lag order for the ACF and PACF.}

\item{main}{Plot title.}
}
\value{
The respective ACF and PACF functions.
}
\description{
Autocorrelation and partial autocorrealtion function.
}
\examples{
ac(rnorm(100), max.lag = 20, main = "Persistence")
}
