% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch_test.R
\name{arch_test}
\alias{arch_test}
\title{Perform Lagrange Multiplier Test for ARCH effect of a time series.}
\usage{
arch_test(data, m = 10)
}
\arguments{
\item{data}{Time series variable.}

\item{m}{Selected AR order.}
}
\value{
ARCH Lagrange Multiplier Statistics.
}
\description{
Perform Lagrange Multiplier Test for ARCH effect of a time series.
}
\examples{
arch_test(rnorm(100))
}
