% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archTest.R
\name{archTest}
\alias{archTest}
\title{Perform Lagrange Multiplier Test for ARCH effect of a time series.}
\usage{
archTest(rtn, m = 10)
}
\arguments{
\item{rtn}{Time series variable.}

\item{m}{Selected AR order.}
}
\value{
ARCH Lagrange Multiplier Statistics.
}
\description{
Perform Lagrange Multiplier Test for ARCH effect of a time series.
}
