% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnd.R
\name{bnd}
\alias{bnd}
\title{Beveridge-Nelson decomposition}
\usage{
bnd(data, nlag = 8)
}
\arguments{
\item{data}{A vector of first-order integrated numeric values}

\item{nlag}{A numberic scalar for the lag-order of the autoregressive (cyclical) part}
}
\value{
Values for the stochastic trend and the stationary cycle
}
\description{
Beveridge-Nelson decomposition
}
\examples{
bnd(rnorm(100))
}
