% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Cross correlation function that considers the relationship between two variables.}
\usage{
cc(x, y, max_lag = NULL, main_title = NULL, output = NULL)
}
\arguments{
\item{x}{A numeric, time series, or xts variable.}

\item{y}{A numeric, time series, or xts variable.}

\item{max_lag}{A number that represents the maximum lag order for the ACF and PACF.}

\item{main_title}{Optional plot title.}

\item{output}{Return data or not.}
}
\value{
The respective ACF and PACF functions.
}
\description{
Cross correlation function that considers the relationship between two variables.
}
\examples{
cc(rnorm(100), rnorm(100))
}
