% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw.R
\name{cw}
\alias{cw}
\title{Clark-West (2007) approximate normality tests for equal predictive accuracy in nested models.}
\usage{
cw(e.m1, e.m2, yf.m1, yf.m2)
}
\arguments{
\item{e.m1}{Errors from model 1.}

\item{e.m2}{Errors from model 2.}

\item{yf.m1}{Forecasts for model 1.}

\item{yf.m2}{Forecasts for model 2.}
}
\value{
Clark-West statistics.
}
\description{
Clark-West (2007) approximate normality tests for equal predictive accuracy in nested models.
}
