% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.R
\name{dif}
\alias{dif}
\title{Returns lagged and differenced data where input and output are same length.}
\usage{
dif(x, lag = 1, differences = 1, ...)
}
\arguments{
\item{x}{A numeric, time series, or xts variable.}

\item{lag}{A number that represents the maximum lag order.}

\item{differences}{Number of times data should be differenced.}

\item{...}{Optional input arguments.}
}
\value{
Vector of data that is the sample length as input.
}
\description{
Returns lagged and differenced data where input and output are same length.
}
\examples{
dif(rnorm(100), 1, 1)
}
