% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Egarch.R
\name{Egarch}
\alias{Egarch}
\title{Estimation of an EGARCH(1,1) model. Assume normal innovations.}
\usage{
Egarch(rtn)
}
\arguments{
\item{rtn}{Time series variable.}
}
\value{
EGARCH results.
}
\description{
Estimation of an EGARCH(1,1) model. Assume normal innovations.
}
