% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch11FIT.R
\name{garch11FIT}
\alias{garch11FIT}
\title{Auxillary function for fitting GARCH model.}
\usage{
garch11FIT(x)
}
\arguments{
\item{x}{time series vector.}
}
\value{
GARCH parameter estimates.
}
\description{
Auxillary function for fitting GARCH model.
}
