% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchM.R
\name{garchM}
\alias{garchM}
\title{Estimation of a Gaussian GARCH-in-Mean with GARCH(1,1) model.}
\usage{
garchM(rtn, type = 1)
}
\arguments{
\item{rtn}{Time series variable.}

\item{type}{1 for Variance-in-mean, 2 for volatility-in-mean, and 3 for log(variance)-in-mean.}
}
\value{
GARCH-M results.
}
\description{
Estimation of a Gaussian GARCH-in-Mean with GARCH(1,1) model.
}
