% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glkM.R
\name{glkM}
\alias{glkM}
\title{Auxillary function for reporting of GARCH-M results.}
\usage{
glkM(pars)
}
\arguments{
\item{pars}{initial results from model.}
}
\value{
GARCH results.
}
\description{
Auxillary function for reporting of GARCH-M results.
}
