% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glkn.R
\name{glkn}
\alias{glkn}
\title{Auxillary function for reporting of NGARCH results.}
\usage{
glkn(par)
}
\arguments{
\item{par}{initial results from model.}
}
\value{
GARCH results.
}
\description{
Auxillary function for reporting of NGARCH results.
}
