% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gts_ur.R
\name{gts_ur}
\alias{gts_ur}
\title{General-to-Specific application of Dickey-Fuller (1981) Test.}
\usage{
gts_ur(series)
}
\arguments{
\item{series}{A vector of numeric values.}
}
\value{
Summary of the results of the various tests.
}
\description{
General-to-Specific application of Dickey-Fuller (1981) Test.
}
\examples{
gts_ur(rnorm(100))
}
