% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Igarch.R
\name{Igarch}
\alias{Igarch}
\title{Estimation of a Gaussian IGARCH(1,1) model.}
\usage{
Igarch(rtn, include.mean = F, volcnt = F)
}
\arguments{
\item{rtn}{Time series variable.}

\item{include.mean}{flag for the constant in the mean equation.}

\item{volcnt}{flag for the constant term of the volatility equation.}
}
\value{
Gaussian IGARCH(1,1) results.
}
\description{
Estimation of a Gaussian IGARCH(1,1) model.
}
