% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadlag.R
\name{leadlag}
\alias{leadlag}
\title{Plot leading and lagging correlations}
\usage{
leadlag(x1, x2, nlag)
}
\arguments{
\item{x1}{A vector of numeric values}

\item{x2}{A vector of numeric values}

\item{nlag}{A numberic scalar for the lag-order of the autoregressive (cyclical) part}
}
\value{
Values for the stochastic trend and the stationary cycle
}
\description{
Plot leading and lagging correlations
}
