% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ngarch.R
\name{Ngarch}
\alias{Ngarch}
\title{Estimation of a non-symmertic GARCH that takes the form NGARCH(1,1) with normal innovations.}
\usage{
Ngarch(rtn)
}
\arguments{
\item{rtn}{Time series variable.}
}
\value{
non-symmertic GARCH results.
}
\description{
Estimation of a non-symmertic GARCH that takes the form NGARCH(1,1) with normal innovations.
}
