% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw.R
\name{nw}
\alias{nw}
\title{Newey-West HAC covariance estimator.}
\usage{
nw(y, qn)
}
\arguments{
\item{y}{A T*k vector of numeric values.}

\item{qn}{A numeric scalar for the truncation lag.}
}
\value{
Newey-West HAC covariance estimator (as derived in Hayashi).
}
\description{
Newey-West HAC covariance estimator.
}
\examples{
nw(rnorm(100), 1)
}
