% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResiVol.R
\name{ResiVol}
\alias{ResiVol}
\title{Auxillary function for reporting of GARCH-M results.}
\usage{
ResiVol(pars)
}
\arguments{
\item{pars}{initial results from model.}
}
\value{
GARCH results.
}
\description{
Auxillary function for reporting of GARCH-M results.
}
