% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round numbers following normal convention.}
\usage{
round2(x, n)
}
\arguments{
\item{x}{A numeric scalar.}

\item{n}{The order for the rounding.}
}
\value{
The result that would be equivalent to Matlab.
}
\description{
Round numbers following normal convention.
}
