% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{ac}
\alias{ac}
\title{Autocorrelation and partial autocorrelation function.}
\usage{
ac(data, max_lag = NULL, main_title = NULL, output = NULL)
}
\arguments{
\item{data}{A numeric, time series, or xts variable.}

\item{max_lag}{A number that represents the maximum lag order for the ACF and PACF.}

\item{main_title}{Optional plot title.}

\item{output}{Return data or not.}
}
\value{
The respective ACF and PACF functions.
}
\description{
Autocorrelation and partial autocorrelation function.
}
\examples{
ac(rnorm(100))
}
