% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSM.R
\name{TSM}
\alias{TSM}
\title{The Smith Method}
\usage{
TSM(x, method = "spearman", corr = seq(0.1, 0.7, by = 0.1), k = 4, verbose = F)
}
\arguments{
\item{x}{Path to the input file}

\item{method}{A Character either \code{pearson} (default) or \code{spearman}, which is the same paramter \code{method} for \code{cor()}.}

\item{corr}{A numeric vector for the thresholds of correlation coefficients.}

\item{verbose}{Boolean}
}
\value{
a data.table (default) or a list of data.table (verbose=T)
}
\description{
This function TSM (aka. The Smith Method) selects a desired number of features (4 by default) by purposefully dropping highly correlated features, \emph{i.e,} picking up a set of representative features that can best explain the binary outcomes. In a plain language, it works like the follwoing: The first representative feature is the one that shows the highest AUC (Area Under the ROC Curve) out of all the features. The next representative feature is the one that shows the highest AUC out of the remaing features after dropping highly correlated features with the first representative feature. The third, the fourth, and so on, represenative feature will be picked up as the same way the 2nd is picked up.
}
\examples{
input=read.csv(system.file("extdata","demo_input.csv",package="TSM")) # read the example input 
TSM(x=input) # run TSM with default parameters

TSM(x=input, corr=c(0.4, 0.5)) # two correlation coefficients only 

TSM(x=input, method="pearson") # pearson method for cor()
}
